Unconditionally await previous promise
Previously, the implementation only waited for the previous line if it
wanted to print anything. This can result in out-of-order logging when
there is a line without result between two lines with results, i.e.:

1. Slow result (wantLine=true)
2. No result (wantLine=false)
3. Fast result (wantLine=true)

In this case, the order of printed lines would be 3,1 instead of 1,3.
Moreover, because the `main` function exits when the last promise
resolves, this may result in the last few lines getting lost.

Fix this by chaining the promises unconditionally.